# src/lambda/binService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de rutas y emergencias",
        "title": "IntegracionDeAplicaciones2 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/cron/update-routes": {
            "post": {
                "description": "Calcula y almacena en caché las rutas óptimas para todas las zonas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron"
                ],
                "summary": "Actualizar todas las rutas en caché",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Clave de API para cron jobs",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de la actualización",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/personas/emails": {
            "get": {
                "description": "Devuelve todas las personas con sus emails asignados para saber qué email usar en cada ruta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener listado de personas con emails",
                "responses": {
                    "200": {
                        "description": "Listado de personas con emails",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/personas/regenerate": {
            "post": {
                "description": "Regenera todas las personas con sus emails desde las rutas existentes en cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Personas"
                ],
                "summary": "Regenerar personas desde rutas",
                "responses": {
                    "200": {
                        "description": "Personas regeneradas exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error regenerando personas",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/routes/neighborhood/{neighborhood}": {
            "get": {
                "description": "Devuelve todas las rutas simplificadas guardadas en caché para un barrio específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rutas"
                ],
                "summary": "Obtener rutas de un barrio",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del barrio",
                        "name": "neighborhood",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de rutas del barrio",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.SimplifiedRoute"
                            }
                        }
                    },
                    "400": {
                        "description": "Neighborhood ID inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron rutas",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/routes/neighborhood/{neighborhood}/route/{routeNumber}": {
            "get": {
                "description": "Devuelve una ruta simplificada específica desde caché",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rutas"
                ],
                "summary": "Obtener ruta específica",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del barrio",
                        "name": "neighborhood",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Número de ruta",
                        "name": "routeNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ruta solicitada",
                        "schema": {
                            "$ref": "#/definitions/services.SimplifiedRoute"
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Ruta no encontrada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.Coordinate": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        },
...
        "services.SimplifiedRoute": {
            "type": "object",
            "properties": {
                "bins_coords": {
                    "description": "Coordenadas de tachos en orden",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.Coordinate"
                    }
                },
                "neighborhood": {
                    "type": "integer"
                },
                "path_coords": {
                    "description": "Coordenadas del path completo",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.Coordinate"
                    }
                },
                "route_id": {
                    "type": "string"
                },
                "total_bins": {
                    "type": "integer"
                },
                "total_distance_km": {
                    "type": "number"
                },
                "truck_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}